Separate dependency installation, use pip through built-in module
